+++ src/test/canister/build.rs
@@ -1,6 +1,6 @@
 fn main() -> std::io::Result<()> {
     // init design dependencies
-    mimic::orm::base::init();
+    mimic::init();
     test_schema::init();
 
     mimic::mimic_build!("test");