+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.2.0]
 - all changes that make mimic compatible with the Dragginz repo v0.2
 - split EntityDynamic into EntityDyn and NodeDyn
+- mimic is now just two crates, as everything else has been incorporated into the main crate
 
 ## [0.1.9]
 - we need to store the Path of an entity inside the db, so we can tell what data it's